requirements.txt
Django==1.11.4
celery==4.1.0
djangorestframework==3.6.3
psycopg2==2.7.3
redis==2.10.6
beautifulsoup4==4.6.0
lxml==3.8.0
urllib3==1.22